#15: add article page and abstract article list into it's own component
