+++ src/pages/index.js
@@ -2,12 +2,9 @@ import React from 'react'
 import styled from 'styled-components'
 import Helmet from 'react-helmet'
 import PropTypes from 'prop-types'
-import Link from 'gatsby-link';
-import moment from 'moment'
 import marked from 'marked'
-import { ArticleHeader, ArticleTitle, ArticleDate } from '../components/Article';
-import Block from '../components/Block';
 import ProjectList from '../components/Project/List';
+import ArticleList from "../components/Article/List";
 
 const TopSection = styled.div`
   display: flex;
@@ -70,40 +67,7 @@ export default class IndexPage extends React.Component {
 
   static propTypes = {
     data: PropTypes.object.isRequired
-  }
-
-  renderArticles() {
-    const { edges: articles } = this.props.data.articles;
-    const Articles = [];
-    articles.forEach((article, key) => {
-      const {
-        title,
-        content: { content },
-        createdAt,
-        slug
-      } = article.node;
-      const date = {
-        day: moment(createdAt).format('D'),
-        month: moment(createdAt).format('MMM'),
-        monthNumber: moment(createdAt).format('MM'),
-        year: moment(createdAt).format('YYYY')
-      };
-      const fullSlug = `/article/${date.year}/${date.monthNumber}/${slug}`;
-      Articles.push(
-        <div key={key}>
-          <ArticleHeader>
-            <ArticleDate>
-              <Block>{date.day}</Block>
-              <Block>{date.month}</Block>
-            </ArticleDate>
-            <ArticleTitle><Link to={fullSlug}>{title}</Link></ArticleTitle>
-          </ArticleHeader>
-          <p dangerouslySetInnerHTML={{ __html: marked(content) }} />
-        </div>
-      );
-    });
-    return Articles;
-  }
+  };
 
   render() {
     const {
@@ -116,6 +80,9 @@ export default class IndexPage extends React.Component {
       },
       projects: {
         edges: projects
+      },
+      articles: {
+        edges: articles
       }
     } = this.props.data;
     return (
@@ -135,7 +102,7 @@ export default class IndexPage extends React.Component {
                 <SectionH2>Articles</SectionH2>
                 <SectionHeaderTagline>Assortment of coding, design, and life</SectionHeaderTagline>
               </SectionHeader>
-              {this.renderArticles()}
+              <ArticleList articles={articles} />
             </Subsection>
             <Subsection>
               <SectionHeader>
